+++ src/libs/tgfclient/guicheckbox.cpp
@@ -99,14 +99,14 @@ GfuiCheckboxCreate(void *scr, int font, int x, int y, int imagewidth, int imageh
 		GfuiGrButtonCreate(scr, "data/img/checked.png", "data/img/checked.png",
 						   "data/img/checked.png", "data/img/checked.png",
 						   x, y, imagewidth, imageheight, GFUI_MIRROR_NONE, false, GFUI_MOUSE_UP,
-						   (void*)(object->id), gfuiChecked,
+						   (void*)(long)(object->id), gfuiChecked,
 						   userDataOnFocus, onFocus, onFocusLost);
 
 	checkbox->uncheckId =
 		GfuiGrButtonCreate(scr, "data/img/unchecked.png", "data/img/unchecked.png",
 						   "data/img/unchecked.png", "data/img/unchecked.png",
 						   x, y, imagewidth, imageheight, GFUI_MIRROR_NONE, false, GFUI_MOUSE_UP,
-						   (void*)(object->id), gfuiUnchecked, 0, 0, 0);
+						   (void*)(long)(object->id), gfuiUnchecked, 0, 0, 0);
 
 	// Compute total height (text or buttons)
 	tGfuiGrButton* pCheckedBut = &(gfuiGetObject(scr, checkbox->checkId)->u.grbutton);
