+++ src/libs/tgfclient/guicombobox.cpp
@@ -142,13 +142,13 @@ GfuiComboboxCreate(void *scr, int font, int x, int y, int width,
 					 "data/img/arrow-left-disabled.png", "data/img/arrow-left.png",
 					 "data/img/arrow-left-focused.png", "data/img/arrow-left-pushed.png",
 					 x, y, arrowsWidth, arrowsHeight, GFUI_MIRROR_NONE,
-					 GFUI_MOUSE_UP,	 (void*)(object->id), gfuiLeftArrow, 0, 0, 0);
+					 GFUI_MOUSE_UP,	 (void*)(long)(object->id), gfuiLeftArrow, 0, 0, 0);
 	gfuiGrButtonInit(&combobox->rightButton,
 					 "data/img/arrow-right-disabled.png", "data/img/arrow-right.png",
 					 "data/img/arrow-right-focused.png", "data/img/arrow-right-pushed.png",
 					 x + width - combobox->leftButton.width, y,
 					 arrowsWidth, arrowsHeight, GFUI_MIRROR_NONE,
-					 GFUI_MOUSE_UP, (void*)(object->id), gfuiRightArrow, 0, 0, 0);
+					 GFUI_MOUSE_UP, (void*)(long)(object->id), gfuiRightArrow, 0, 0, 0);
 
 	// Compute total height (text or buttons)
 	int height = gfuiFont[font]->getHeight();
